+++ lib/morphable_shape.dart
@@ -4,8 +4,9 @@ export 'package:dimension/dimension.dart';
 export 'package:flutter_class_parser/parse_json.dart';
 export 'package:flutter_class_parser/to_json.dart';
 
-export 'animated_shadowd_shape.dart';
+export 'animated_decorated_shadowd_shape.dart';
 export 'border_paths.dart';
+export 'decorated_shadowed_shape.dart';
 export 'dynamic_border_radius.dart';
 export 'dynamic_border_side.dart';
 export 'dynamic_path.dart';
@@ -13,7 +14,6 @@ export 'dynamic_rectangle_styles.dart';
 export 'morphable_shape_border.dart';
 export 'parse_json.dart';
 export 'preset_shape_map.dart';
-export 'shadowed_shape.dart';
 export 'shape.dart';
 export 'shape_shadow.dart';
 export 'shapes/arc.dart';